java/155_MinStack/src/test/java/com/kdnakt/leetcode/MinStackTest.java
package com.kdnakt.leetcode;

import static org.junit.jupiter.api.Assertions.*;

import org.junit.jupiter.api.Test;

public class MinStackTest {

    @Test
    public void testExample1GetMin() {
        MinStack minStack = new MinStack();
        minStack.push(-2);
        minStack.push(0);
        minStack.push(-3);
        assertEquals(-3, minStack.getMin());
    }

    @Test
    public void testExample1PopAndTop() {
        MinStack minStack = new MinStack();
        minStack.push(-2);
        minStack.push(0);
        minStack.push(-3);
        minStack.getMin();
        minStack.pop();
        assertEquals(0, minStack.top());
    }

    @Test
    public void testExample1GetMin2() {
        MinStack minStack = new MinStack();
        minStack.push(-2);
        minStack.push(0);
        minStack.push(-3);
        minStack.getMin();
        minStack.pop();
        minStack.top();
        assertEquals(-2, minStack.getMin());
    }
}
